+++ crates/koja-ir-eval/src/intrinsics/mod.rs
@@ -32,6 +32,7 @@ mod numeric;
 mod parse;
 mod print;
 mod process;
+mod runtime_block;
 mod set;
 mod socket;
 mod string;
@@ -60,7 +61,7 @@ pub(crate) async fn dispatch<R: CallResolver>(
     resolver: &R,
 ) -> Result<Value, RuntimeError> {
     match *id {
-        IRIntrinsicId::Binary(method) => binary::binary(method, function, args),
+        IRIntrinsicId::Binary(method) => binary::binary(method, function, args, resolver),
         IRIntrinsicId::Bits(method) => binary::bits(method, function, args),
         IRIntrinsicId::Bitwise { ty, op } => bitwise::dispatch(ty, op, args),
         IRIntrinsicId::CPtr(method) => cptr::dispatch(method, function, args),
@@ -83,6 +84,7 @@ pub(crate) async fn dispatch<R: CallResolver>(
         IRIntrinsicId::ReplyTo(method) => {
             process::reply_to_dispatch(method, function, args, resolver).await
         }
+        IRIntrinsicId::RuntimeBlock(method) => runtime_block::dispatch(method, args),
         IRIntrinsicId::Set(method) => set::dispatch(method, function, args),
         IRIntrinsicId::Socket(method) => socket::dispatch(method, function, args, resolver).await,
         IRIntrinsicId::String(method) => string::dispatch(method, function, args, resolver),